requirements.txt
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Farama-Notifications==0.0.4
fonttools==4.58.4
graphviz==0.21
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
kiwisolver==1.4.8
matplotlib==3.10.3
neat-python==0.92
networkx==3.5
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.14.0
